Fixes: #144. Updated client and server to better handle 64-bit cash values.

+++ client/src/main/java/au/com/codeka/Cash.java
@@ -2,7 +2,7 @@ package au.com.codeka;
 
 
 public class Cash {
-    public static String format(float cash) {
-        return "$"+NumberFormatter.format((int) Math.floor(cash));
+    public static String format(double cash) {
+        return "$"+NumberFormatter.format((long) Math.floor(cash));
     }
 }
